Add files via upload
A script that scrapes data from the internet to calculate the intrinsic value of a stock using discounted cash flow model. The values used to calculate are the mean of free cash flow growth rate and the lowest end of confidence interval at 80% of free cash flow growth rate. Confidence intervals and t values were used to calculate the range of free cash flow growth rate.